Time: 45 ms (84.35%) | Memory: 79.2 MB (26.69%) - LeetSync

+++ 1675-magnetic-force-between-two-balls/magnetic-force-between-two-balls.java
@@ -2,13 +2,13 @@ class Solution {
     public int maxDistance(int[] position, int m) {
         Arrays.sort(position);
         int n=position.length;
-        long ans=-1;
-        long low=0;
-        long high=position[n-1]-position[0];
+        int ans=-1;
+        int low=0;
+        int high=position[n-1]-position[0];
         while(low<=high){
-            long mid=(high+low)/2;
-            long ball=1;
-            long last=position[0];
+            int mid=(high+low)/2;
+            int ball=1;
+            int last=position[0];
             for(int i=1;i<n;i++){
                 if(position[i]-last>=mid){
                     ball++;
@@ -23,6 +23,6 @@ class Solution {
                     high=mid-1;
                 }
         }
-        return (int)ans;
+        return ans;
     }
 }
